+++ internal/engine/netx/dialer/logging_test.go
@@ -2,7 +2,6 @@ package dialer_test
 
 import (
 	"context"
-	"crypto/tls"
 	"errors"
 	"io"
 	"testing"
@@ -22,21 +21,5 @@ func TestLoggingDialerFailure(t *testing.T) {
 	}
 	if conn != nil {
 		t.Fatal("expected nil conn here")
-	}
-}
-
-func TestLoggingTLSHandshakerFailure(t *testing.T) {
-	h := dialer.LoggingTLSHandshaker{
-		TLSHandshaker: dialer.EOFTLSHandshaker{},
-		Logger:        log.Log,
-	}
-	tlsconn, _, err := h.Handshake(context.Background(), dialer.EOFConn{}, &tls.Config{
-		ServerName: "www.google.com",
-	})
-	if !errors.Is(err, io.EOF) {
-		t.Fatal("not the error we expected")
-	}
-	if tlsconn != nil {
-		t.Fatal("expected nil tlsconn here")
 	}
 }
