+++ internal/engine/netx/tlsdialer/eof_test.go
@@ -1,7 +1,8 @@
-package dialer
+package tlsdialer
 
 import (
 	"context"
+	"crypto/tls"
 	"io"
 	"net"
 	"time"
@@ -67,4 +68,13 @@ func (EOFAddr) Network() string {
 
 func (EOFAddr) String() string {
 	return "127.0.0.1:1234"
+}
+
+type EOFTLSHandshaker struct{}
+
+func (EOFTLSHandshaker) Handshake(
+	ctx context.Context, conn net.Conn, config *tls.Config,
+) (net.Conn, tls.ConnectionState, error) {
+	time.Sleep(10 * time.Microsecond)
+	return nil, tls.ConnectionState{}, io.EOF
 }
